requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
python-dotenv==0.21.1
youtube-transcript-api==0.6.1
openai==1.3.0
python-multipart==0.0.6
jinja2==3.1.2
requests==2.31.0 